refactor: moved header & snooze option to their own files
added check before toUpperCase function
