+++ apps/app/components/notifications/index.ts
@@ -1,3 +1,4 @@
 export * from "./notification-card";
 export * from "./notification-popover";
 export * from "./select-snooze-till-modal";
+export * from "./notification-header";
