+++ apps/app/components/notifications/notification-card.tsx
@@ -21,6 +21,8 @@ import {
 
 // type
 import type { IUserNotification } from "types";
+// constants
+import { snoozeOptions } from "constants/notification";
 
 type NotificationCardProps = {
   notification: IUserNotification;
@@ -29,33 +31,6 @@ type NotificationCardProps = {
   setSelectedNotificationForSnooze: (notificationId: string) => void;
   markSnoozeNotification: (notificationId: string, dateTime?: Date | undefined) => Promise<void>;
 };
-
-const snoozeOptions = [
-  {
-    label: "1 day",
-    value: new Date(new Date().getTime() + 24 * 60 * 60 * 1000),
-  },
-  {
-    label: "3 days",
-    value: new Date(new Date().getTime() + 3 * 24 * 60 * 60 * 1000),
-  },
-  {
-    label: "5 days",
-    value: new Date(new Date().getTime() + 5 * 24 * 60 * 60 * 1000),
-  },
-  {
-    label: "1 week",
-    value: new Date(new Date().getTime() + 7 * 24 * 60 * 60 * 1000),
-  },
-  {
-    label: "2 weeks",
-    value: new Date(new Date().getTime() + 14 * 24 * 60 * 60 * 1000),
-  },
-  {
-    label: "Custom",
-    value: null,
-  },
-];
 
 export const NotificationCard: React.FC<NotificationCardProps> = (props) => {
   const {
@@ -101,7 +76,11 @@ export const NotificationCard: React.FC<NotificationCardProps> = (props) => {
         ) : (
           <div className="w-12 h-12 bg-custom-background-80 rounded-full flex justify-center items-center">
             <span className="text-custom-text-100 font-medium text-lg">
-              {notification.triggered_by_details.first_name[0].toUpperCase()}
+              {notification.triggered_by_details.first_name?.[0] ? (
+                notification.triggered_by_details.first_name?.[0]?.toUpperCase()
+              ) : (
+                <Icon iconName="person" className="h-6 w-6" />
+              )}
             </span>
           </div>
         )}
